pyproject.toml
[project]
name = "trading-brain"
version = "0.1.0"
description = "Self-adjusting AI trading brain with three-brain architecture"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-telegram-bot>=22.0",
    "apscheduler>=3.11",
    "aiosqlite>=0.20.0",
    "httpx>=0.27",
    "pandas>=2.0",
    "numpy>=1.26",
    "ta>=0.11",
    "python-dotenv>=1.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
trading-brain = "src.main:run"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
